requirements.txt
Flask>=2.3.0
tensorflow>=2.15.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
requests>=2.31.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
Werkzeug>=2.3.0
